src/main/webapp/app/entities/fundraising-idea/delete/fundraising-idea-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { IFundraisingIdea } from '../fundraising-idea.model';
import { FundraisingIdeaService } from '../service/fundraising-idea.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  templateUrl: './fundraising-idea-delete-dialog.component.html',
})
export class FundraisingIdeaDeleteDialogComponent {
  fundraisingIdea?: IFundraisingIdea;

  constructor(protected fundraisingIdeaService: FundraisingIdeaService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.fundraisingIdeaService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
